Add an action for adding members to a chat room
Even though django-rest-framework supports a Ruby-on-Rails style
of updating existing resources by issuing a PATCH or PUT request,
such updates are unsafe and can cause race-conditions to lose
some state. The implementation of this action isn't fully RESTful,
but neither is the PATCH-style used here, so it is still a good
choice for solving a possible race condition situation.