Refactor development requirements file
Instead of having the development requirements file list all the
packages which are required in production, adapt the script to install
both packages from ``requirements.txt`` as well as
``requirements-dev.txt``.  That way, the development requirements file
can be stated only in terms of *extra* packages which are useful
for a development environment.