Implement a GCM notifier class
The class exposes a convenient interface for sending out notifications
when a new message is created. Instances of the class take care to
automatically gather the list of users which need to be notified
about the message (the ones belonging to the chat room in which the
message was posted) and sending out a request to the GCM service.

The ``notify`` method is completely synchronous and care should be
taken not to use the ``GcmNotifier`` within a time-sensitive context
such as an HTTP request/response cycle. The notifer should be used
in an async manner, e.g. queuing the notification as a celery task.

There are settings values for the project which can disable sending
GCM-based notifications, if necessary.