Set up Celery using RabbitMQ as the Broker
Sets up all the requirements to use celery as the asnyc task manager.

It makes sure to install RabbitMQ to the Vagrant box so that celery
integration with a real message queue can be tested. However, the
default development settings specify that all celery tasks run
synchronously. This is done in order to prevent issues for developers
who do not wish to develop any features depending on the async
execution of tasks.