Implement a SystemMessage model
The model represents messages emitted by the system into chat rooms.
These messages are to be various status messages, such as a notification
that a user has left or joined the chat room.

The class makes sure the system is extensible enough to support any
kind of system-generated message in the future.